app/requirements.txt
requests
python-dotenv
web3